# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "i10 API",
    "version": "0.0.0",
    "description": "Transactional email. The document is generated from the same Zod schemas the server validates with and the SDKs are built from, so it cannot drift from the implementation."
  },
  "servers": [
    {
      "url": "https://api.i10.tech",
      "description": "Production"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Your i10 API key. Keys are prefixed `i10_live_` or `i10_test_` so they are recognisable in your own logs and greppable in a leak scan."
      }
    },
    "schemas": {
      "SendEmailResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "id"
        ]
      },
      "Error": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "integer"
          },
          "name": {
            "type": "string",
            "enum": [
              "validation_error",
              "missing_required_field",
              "invalid_access",
              "not_found",
              "method_not_allowed",
              "rate_limit_exceeded",
              "daily_quota_exceeded",
              "invalid_from_address",
              "invalid_to_address",
              "domain_not_verified",
              "idempotency_conflict",
              "internal_server_error"
            ]
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "statusCode",
          "name",
          "message"
        ]
      },
      "SendEmail": {
        "type": "object",
        "properties": {
          "from": {
            "type": "string",
            "minLength": 3,
            "maxLength": 320
          },
          "to": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 3,
                "maxLength": 320
              },
              {
                "type": "array",
                "items": {
                  "type": "string",
                  "minLength": 3,
                  "maxLength": 320
                },
                "minItems": 1,
                "maxItems": 50
              }
            ]
          },
          "subject": {
            "type": "string"
          },
          "bcc": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 3,
                "maxLength": 320
              },
              {
                "type": "array",
                "items": {
                  "type": "string",
                  "minLength": 3,
                  "maxLength": 320
                },
                "minItems": 1,
                "maxItems": 50
              }
            ]
          },
          "cc": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 3,
                "maxLength": 320
              },
              {
                "type": "array",
                "items": {
                  "type": "string",
                  "minLength": 3,
                  "maxLength": 320
                },
                "minItems": 1,
                "maxItems": 50
              }
            ]
          },
          "reply_to": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 3,
                "maxLength": 320
              },
              {
                "type": "array",
                "items": {
                  "type": "string",
                  "minLength": 3,
                  "maxLength": 320
                },
                "minItems": 1,
                "maxItems": 50
              }
            ]
          },
          "html": {
            "type": "string"
          },
          "text": {
            "type": "string"
          },
          "headers": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          },
          "attachments": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "filename": {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 255,
                  "pattern": "^[^\\r\\n\"]+$"
                },
                "content": {
                  "type": "string"
                },
                "path": {
                  "type": "string",
                  "format": "uri"
                },
                "content_type": {
                  "type": "string",
                  "maxLength": 255
                }
              },
              "required": [
                "filename"
              ]
            }
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "pattern": "^[A-Za-z0-9_-]{1,256}$"
                },
                "value": {
                  "type": "string",
                  "pattern": "^[A-Za-z0-9_-]{1,256}$"
                }
              },
              "required": [
                "name",
                "value"
              ]
            }
          },
          "scheduled_at": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "from",
          "to",
          "subject"
        ]
      },
...
      "Email": {
        "type": "object",
        "properties": {
          "object": {
            "type": "string",
            "enum": [
              "email"
            ]
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "from": {
            "type": "string"
          },
          "to": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "cc": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "bcc": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "reply_to": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "subject": {
            "type": "string"
          },
          "html": {
            "type": [
              "string",
              "null"
            ]
          },
          "text": {
            "type": [
              "string",
              "null"
            ]
          },
          "created_at": {
            "type": "string"
          },
          "scheduled_at": {
            "type": [
              "string",
              "null"
            ]
          },
          "last_event": {
            "type": "string",
            "enum": [
              "queued",
              "scheduled",
              "sending",
              "sent",
              "delivered",
              "delivery_delayed",
              "bounced",
              "complained",
              "failed",
              "canceled"
            ]
          }
        },
        "required": [
          "object",
          "id",
          "from",
          "to",
          "cc",
          "bcc",
          "reply_to",
          "subject",
          "html",
          "text",
          "created_at",
          "scheduled_at",
          "last_event"
        ]
      },
      "WebhookEndpointWithSecret": {
        "type": "object",
        "properties": {
          "object": {
            "type": "string",
            "enum": [
              "webhook_endpoint"
            ]
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "url": {
            "type": "string",
            "format": "uri"
          },
          "events": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "email.sent",
                "email.delivered",
                "email.delivery_delayed",
                "email.bounced",
                "email.complained",
                "email.failed"
              ]
            }
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "enabled": {
            "type": "boolean"
          },
          "created_at": {
            "type": "string"
          },
          "secret": {
            "type": "string"
          }
        },
        "required": [
          "object",
          "id",
          "url",
          "events",
          "description",
          "enabled",
          "created_at",
          "secret"
        ]
      },
      "CreateWebhookEndpoint": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string",
            "format": "uri"
          },
          "events": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "email.sent",
                "email.delivered",
                "email.delivery_delayed",
                "email.bounced",
                "email.complained",
                "email.failed"
              ]
            },
            "minItems": 1
          },
          "description": {
            "type": "string",
            "maxLength": 255
          }
        },
        "required": [
          "url",
          "events"
        ]
      },
      "WebhookEndpointList": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "object": {
                  "type": "string",
                  "enum": [
                    "webhook_endpoint"
                  ]
                },
                "id": {
                  "type": "string",
                  "format": "uuid"
                },
                "url": {
                  "type": "string",
                  "format": "uri"
                },
                "events": {
                  "type": "array",
                  "items": {
                    "type": "string",
                    "enum": [
                      "email.sent",
                      "email.delivered",
                      "email.delivery_delayed",
                      "email.bounced",
                      "email.complained",
                      "email.failed"
                    ]
                  }
                },
                "description": {
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "enabled": {
                  "type": "boolean"
                },
                "created_at": {
                  "type": "string"
                }
              },
              "required": [
                "object",
                "id",
                "url",
                "events",
                "description",
                "enabled",
                "created_at"
              ]
            }
          }
        },
        "required": [
          "data"
        ]
      }
    },
...
  },
  "paths": {
    "/emails": {
      "post": {
        "summary": "Send an email",
        "description": "Queues a single message and returns its id immediately. The id is minted in the same transaction as the caller's own work, so a rollback never sends.",
        "tags": [
          "Emails"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendEmail"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The message was accepted.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SendEmailResponse"
                }
              }
            }
          },
          "401": {
            "description": "The API key is missing, malformed, or unknown.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "This Idempotency-Key was already used with a different request body.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "The request body failed validation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Sending too fast, or the daily quota is exhausted.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "501": {
            "description": "The send path is not configured.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/emails/{id}": {
      "get": {
        "summary": "Retrieve an email",
        "description": "The message as it was accepted, and how far it has got. `last_event` is the furthest state reached, so a bounce is visible even though the message was accepted successfully.",
        "tags": [
          "Emails"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "The message.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Email"
                }
              }
            }
          },
          "401": {
            "description": "The API key is missing, malformed, or unknown.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "No such message for this API key's tenant.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "`id` is not a uuid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "501": {
            "description": "The send path is not configured.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/webhook-endpoints": {
      "post": {
        "summary": "Create a webhook endpoint",
        "description": "Registers an https URL to receive events. The signing secret is returned here and never again — store it before you discard the response.",
        "tags": [
          "Webhooks"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWebhookEndpoint"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Created. `secret` is shown only in this response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebhookEndpointWithSecret"
                }
              }
            }
          },
          "401": {
            "description": "The API key is missing, malformed, or unknown.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "The URL or the event list is not acceptable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "501": {
            "description": "Webhook endpoints are not configured.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "List webhook endpoints",
        "tags": [
          "Webhooks"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Every endpoint for this tenant. Secrets are never included.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebhookEndpointList"
                }
              }
            }
          },
          "401": {
            "description": "The API key is missing, malformed, or unknown.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "501": {
            "description": "Webhook endpoints are not configured.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}